Add files via upload
Project Name - Snake Game
Language - C++
IDE - Codeblocks
Actually It's not a big or good project. I was actually trying to make Floppy Bird game by Python. But I've found some bug, for this reason I couldn't finish it and did the snake game.
I included a term called playing status which ensure the player that is he/she playing the game or pause it or died in it. Hope you all like it.
Keep me in your prayer.